Consolidate Daylite test support and split projects tests
Extract the MockTransport, mock_response, and token-state fixtures
duplicated across projects.rs, contacts/api.rs, and auth_flow.rs into a
shared daylite::test_support module.
Derive Default for DayliteSearchInput so tests only spell out relevant
fields.
Move the projects test module into projects/tests.rs.
Remove narration, history, and duplicated constraint comments in the
touched files per the refactoring plan, which is updated to also drop
backlog-ID test markers and deduplicate constraint comments.
