+++ src-tauri/src/integrations/daylite/mod.rs
@@ -6,3 +6,5 @@ pub mod projects;
 #[cfg(test)]
 mod recording_harness;
 pub mod shared;
+#[cfg(test)]
+mod test_support;
